ref: move row context menu ui and logic to separate component
