+++ packages/extension/src/view/devtools/components/cookies/cookiesListing/index.tsx
@@ -23,7 +23,6 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { createPortal } from 'react-dom';
 import { Resizable } from 're-resizable';
 import {
   filterCookiesByFrame,
@@ -37,7 +36,6 @@ import {
   type InfoType,
   type TableColumn,
   type TableFilter,
-  type TableRow,
 } from '@ps-analysis-tool/design-system';
 
 /**
@@ -45,10 +43,8 @@ import {
  */
 import { useCookieStore } from '../../../stateProviders/syncCookieStore';
 import { BLOCKED_REASON_LIST } from '../../../../../constants';
-import { getCurrentTab } from '../../../../../utils/getCurrentTabId';
-import setDomainsInAllowList from './handleAllowList/setDomainsInAllowList';
-import onAllowListClick from './handleAllowList/onAllowListClick';
-import setParentDomain from './handleAllowList/setParentDomain';
+import RowContextMenu from './rowContextMenu';
+import useAllowedList from './useAllowedList';
 
 interface CookiesListingProps {
   setFilteredCookies: React.Dispatch<CookieTableData[]>;
@@ -63,90 +59,23 @@ const CookiesListing = ({ setFilteredCookies }: CookiesListingProps) => {
     getCookiesSetByJavascript,
   } = useCookieStore(({ state, actions }) => ({
     selectedFrame: state.selectedFrame,
-    cookies: state.tabCookies || {},
+    cookies: state.tabCookies,
     tabFrames: state.tabFrames,
     tabUrl: state.tabUrl,
     getCookiesSetByJavascript: actions.getCookiesSetByJavascript,
   }));
 
-  const isIncognito = useRef<boolean>(false);
-
-  useEffect(() => {
-    (async () => {
-      const tabs = await getCurrentTab();
-
-      if (tabs?.length) {
-        isIncognito.current = tabs[0].incognito;
-      }
-    })();
-  }, [tabUrl]);
-
   const [tableData, setTableData] = useState<Record<string, CookieTableData>>(
     {}
   );
 
-  const [domainsInAllowList, setDomainsInAllowListCallback] = useState<
-    Set<string>
-  >(new Set());
-
-  const domainsInAllowListRef = useRef<Set<string>>(new Set());
-
-  useEffect(() => {
-    domainsInAllowListRef.current = domainsInAllowList;
-  }, [domainsInAllowList]);
-
-  useEffect(() => {
-    (async () => {
-      const setter = (list: Set<string>) => {
-        domainsInAllowListRef.current = list;
-      };
-
-      await Promise.all(
-        Object.values(cookies).map(async (cookie) => {
-          if (cookie.parsedCookie?.domain) {
-            const domain = cookie.parsedCookie.domain || '';
-            let isDomainInAllowList = domainsInAllowListRef.current.has(domain);
-
-            if (!isDomainInAllowList) {
-              isDomainInAllowList = [...domainsInAllowListRef.current].some(
-                (storedDomain) => {
-                  // For example xyz.bbc.com and .bbc.com
-                  if (
-                    (domain.endsWith(storedDomain) &&
-                      domain !== storedDomain) ||
-                    `.${domain}` === storedDomain
-                  ) {
-                    return true;
-                  }
-
-                  return false;
-                }
-              );
-            }
-
-            if (isDomainInAllowList) {
-              return;
-            }
-
-            await setDomainsInAllowList(
-              tabUrl || '',
-              isIncognito.current,
-              cookie.parsedCookie.domain,
-              domainsInAllowListRef.current,
-              setter
-            );
-          }
-        })
-      );
-
-      setDomainsInAllowListCallback(domainsInAllowListRef.current);
-    })();
-  }, [cookies, tabUrl]);
+  const { domainsInAllowList, setDomainsInAllowListCallback, isIncognito } =
+    useAllowedList();
 
   useEffect(() => {
     setTableData((prevData) =>
       Object.fromEntries(
-        Object.entries(cookies).map(([key, cookie]) => {
+        Object.entries(cookies || {}).map(([key, cookie]) => {
           const domain = cookie.parsedCookie?.domain || '';
           let isDomainInAllowList = domainsInAllowList.has(domain);
 
@@ -174,118 +103,6 @@ const CookiesListing = ({ setFilteredCookies }: CookiesListingProps) => {
       )
     );
   }, [cookies, domainsInAllowList]);
-
-  const [parentDomain, setParentDomainCallback] = useState<string>('');
-
-  const [contextMenuOpen, setContextMenuOpen] = useState(false);
-  const [columnPosition, setColumnPosition] = useState({
-    x: 0,
-    y: 0,
-  });
-  const [selectedCookie, setSelectedCookie] = useState<CookieTableData | null>(
-    null
-  );
-
-  const handleRightClick = useCallback(
-    async (e: React.MouseEvent<HTMLElement>, { originalData }: TableRow) => {
-      e.preventDefault();
-      const x = e.clientX,
-        y = e.clientY;
-      setColumnPosition({ x, y });
-      document.body.style.overflow = contextMenuOpen ? 'auto' : 'hidden';
-      setContextMenuOpen(!contextMenuOpen);
-      setSelectedCookie(originalData as CookieTableData);
-      await setParentDomain(
-        (originalData as CookieTableData).parsedCookie.domain || '',
-        setParentDomainCallback
-      );
-    },
-    [contextMenuOpen]
-  );
-
-  const [domain, name] = useMemo(
-    () => [
-      selectedCookie?.parsedCookie?.domain || '',
-      selectedCookie?.parsedCookie?.name,
-    ],
-    [selectedCookie]
-  );
-
-  const isDomainInAllowList = useMemo(() => {
-    let _isDomainInAllowList = domainsInAllowList.has(domain);
-
-    if (!_isDomainInAllowList) {
-      _isDomainInAllowList = [...domainsInAllowList].some((storedDomain) => {
-        // For example xyz.bbc.com and .bbc.com
-        if (
-          (domain.endsWith(storedDomain) && domain !== storedDomain) ||
-          `.${domain}` === storedDomain
-        ) {
-          return true;
-        }
-
-        return false;
-      });
-    }
-
-    return _isDomainInAllowList;
-  }, [domain, domainsInAllowList]);
-
-  const handleCopy = useCallback(() => {
-    try {
-      // Need to do this since chrome doesnt allow the clipboard access in extension.
-      const copyFrom = document.createElement('textarea');
-      copyFrom.textContent = `cookie-domain:${domain} cookie-name:${name}`;
-      document.body.appendChild(copyFrom);
-      copyFrom.select();
-      document.execCommand('copy');
-      copyFrom.blur();
-      document.body.removeChild(copyFrom);
-      setContextMenuOpen(false);
-    } catch (error) {
-      //Fail silently
-    }
-  }, [domain, name]);
-
-  const cookieTableRef = useRef<React.ElementRef<typeof CookieTable> | null>(
-    null
-  );
-
-  const handleAllowListClick = useCallback(
-    async (e: React.MouseEvent<HTMLElement>) => {
-      e.stopPropagation();
-
-      cookieTableRef.current?.removeSelectedRow();
-      await onAllowListClick(
-        domain,
-        tabUrl || '',
-        isIncognito.current,
-        domainsInAllowList.has(domain),
-        domainsInAllowList,
-        setDomainsInAllowListCallback
-      );
-      setContextMenuOpen(false);
-    },
-    [domain, domainsInAllowList, tabUrl]
-  );
-
-  const handleAllowListWithParentDomainClick = useCallback(
-    async (e: React.MouseEvent<HTMLElement>) => {
-      e.stopPropagation();
-
-      cookieTableRef.current?.removeSelectedRow();
-      await onAllowListClick(
-        parentDomain,
-        tabUrl || '',
-        isIncognito.current,
-        domainsInAllowList.has(parentDomain),
-        domainsInAllowList,
-        setDomainsInAllowListCallback
-      );
-      setContextMenuOpen(false);
-    },
-    [domainsInAllowList, parentDomain, tabUrl]
-  );
 
   const removeHighlights = useCallback(() => {
     setTableData((prev) =>
@@ -621,6 +438,14 @@ const CookiesListing = ({ setFilteredCookies }: CookiesListingProps) => {
     [getCookiesSetByJavascript]
   );
 
+  const cookieTableRef = useRef<React.ElementRef<typeof CookieTable> | null>(
+    null
+  );
+
+  const rowContextMenuRef = useRef<React.ElementRef<
+    typeof RowContextMenu
+  > | null>(null);
+
   return (
     <div className="w-full h-full flex flex-col">
       <Resizable
@@ -649,62 +474,19 @@ const CookiesListing = ({ setFilteredCookies }: CookiesListingProps) => {
           selectedFrameCookie={selectedFrameCookie}
           setSelectedFrameCookie={setSelectedFrameCookie}
           extraInterfaceToTopBar={extraInterfaceToTopBar}
-          onRowContextMenu={handleRightClick}
+          onRowContextMenu={rowContextMenuRef.current?.onRowContextMenu}
           ref={cookieTableRef}
         />
       </Resizable>
       <CookieDetails selectedFrameCookie={selectedFrameCookie} />
-      <>
-        {domain &&
-          name &&
-          contextMenuOpen &&
-          createPortal(
-            <div className="transition duration-100" data-testid="column-menu">
-              <div
-                className="absolute z-50 text-raisin-black dark:text-bright-gray rounded-md backdrop-blur-2xl p-1.5 mr-2 divide-neutral-300 dark:divide-neutral-500 max-h-[78vh] bg-stone-200 dark:bg-neutral-700 shadow-3xl"
-                style={{
-                  left:
-                    'min( calc( 100vw - 15rem),' + columnPosition.x + 'px )',
-                  top: columnPosition.y + 'px',
-                  border: '0.5px solid rgba(0, 0, 0, 0.20)',
-                }}
-              >
-                <button
-                  onClick={handleCopy}
-                  className="w-full text-xs rounded px-1 py-[3px] flex items-center hover:bg-royal-blue hover:text-white cursor-default"
-                >
-                  <span>Copy network filter string</span>
-                </button>
-
-                {isDomainInAllowList && parentDomain ? (
-                  <button
-                    onClick={handleAllowListWithParentDomainClick}
-                    className="w-full text-xs rounded px-1 py-[3px] flex items-center hover:bg-royal-blue hover:text-white cursor-default"
-                  >
-                    <span>Remove `{parentDomain}` from allow list</span>
-                  </button>
-                ) : (
-                  <button
-                    onClick={handleAllowListClick}
-                    className="w-full text-xs rounded px-1 py-[3px] flex items-center hover:bg-royal-blue hover:text-white cursor-default"
-                  >
-                    <span id="allow-list-option">
-                      {isDomainInAllowList
-                        ? 'Remove domain from allow list'
-                        : 'Add domain to allow list'}
-                    </span>
-                  </button>
-                )}
-              </div>
-              <div
-                data-testid="column-menu-overlay"
-                onClick={() => setContextMenuOpen(false)}
-                className="absolute w-screen h-screen z-10 top-0 left-0"
-              />
-            </div>,
-            document.body
-          )}
-      </>
+      <RowContextMenu
+        domainsInAllowList={domainsInAllowList}
+        isIncognito={isIncognito}
+        tabUrl={tabUrl || ''}
+        setDomainsInAllowListCallback={setDomainsInAllowListCallback}
+        removeSelectedRow={cookieTableRef.current?.removeSelectedRow}
+        ref={rowContextMenuRef}
+      />
     </div>
   );
 };
